Rubik's Cube/requirements.txt
pygame>=1.9.3
PyOpenGL>=3.1.0